fix: update location of decryption
