+++ py_hamt/__init__.py
@@ -1,5 +1,6 @@
 from .hamt import HAMT, blake3_hashfn
 from .store import Store, DictStore, IPFSStore
+from .zarr_encryption_transformer import create_zarr_encryption_transformers
 
 __all__ = [
     "HAMT",
@@ -7,4 +8,5 @@ __all__ = [
     "Store",
     "DictStore",
     "IPFSStore",
+    "create_zarr_encryption_transformers",
 ]
