+++ pyproject.toml
@@ -9,6 +9,7 @@ dependencies = [
     "msgspec>=0.18.6",
     "multiformats[full]>=0.3.1.post4",
     "requests>=2.32.3",
+    "pycryptodome",
 ]
 
 [build-system]
@@ -28,9 +29,4 @@ dev = [
     "numpy>=2.1.3",
     "xarray>=2024.11.0",
     "zarr>=2.18.3",
-]
-
-[tool.uv]
-dev-dependencies = [
-    "pycryptodome",
 ]
